Refactor database app into modules
