+++ public/db/app/constants.js
@@ -1,0 +1,1 @@
+export const PAGE_SIZE = 12;
