+++ public/db/app/elements.js
@@ -1,0 +1,27 @@
+export const elements = {
+  searchForm: document.getElementById('searchForm'),
+  searchInput: document.getElementById('searchInput'),
+  typeFilter: document.getElementById('typeFilter'),
+  rarityFilter: document.getElementById('rarityFilter'),
+  resultInfo: document.getElementById('resultInfo'),
+  cardGrid: document.getElementById('cardGrid'),
+  loadMoreButton: document.getElementById('loadMoreButton'),
+  detailView: document.getElementById('detailView'),
+  detailMedia: document.getElementById('detailMedia'),
+  detailImage: document.getElementById('detailImage'),
+  detailFallback: document.getElementById('detailFallback'),
+  detailTitle: document.getElementById('detailTitle'),
+  detailSubtitle: document.getElementById('detailSubtitle'),
+  detailStars: document.getElementById('detailStars'),
+  detailRarity: document.getElementById('detailRarity'),
+  detailType: document.getElementById('detailType'),
+  detailReleaseRelative: document.getElementById('detailReleaseRelative'),
+  detailReleaseAbsolute: document.getElementById('detailReleaseAbsolute'),
+  detailDescriptionSection: document.getElementById('detailDescriptionSection'),
+  detailDescription: document.getElementById('detailDescription'),
+  detailPropertiesSection: document.getElementById('detailPropertiesSection'),
+  detailProperties: document.getElementById('detailProperties'),
+  detailEnchantmentsSection: document.getElementById('detailEnchantmentsSection'),
+  detailEnchantments: document.getElementById('detailEnchantments'),
+  detailStatus: document.getElementById('detailStatus')
+};
