+++ public/db/app/index.js
@@ -1,0 +1,5 @@
+import { initializeApp } from './initialize.js';
+
+initializeApp();
+
+export { initializeApp };
