+++ public/db/index.html
@@ -123,6 +123,6 @@
       </article>
     </aside>
 
-    <script type="module" src="./app.js"></script>
+    <script type="module" src="./app/index.js"></script>
   </body>
 </html>
